+++ src/ClassLibraryNet45/Class1.cs
@@ -1,22 +1,23 @@
 ﻿using System;
 using System.ComponentModel.Composition;
+using System.Globalization;
 using System.Windows;
 using System.Windows.Controls;
-using KsWare.Presentation.Interfaces.Plugins.TemplateConverter;
+using KsWare.Presentation.Interfaces.Plugins.ResourceConverter;
 
 namespace ClassLibraryNet45
 {
-	[Export(typeof(ITemplateConverterPlugin))]
-	[TemplateConverterPluginExportMetadata("test/test")]
-	public class Class1 : ITemplateConverterPlugin {
+	[Export(typeof(IResourceConverterPlugin))]
+	[ResourceConverterPluginExportMetadata("test/test")]
+	public class Class1 : IResourceConverterPlugin {
 
-	    public DataTemplate CreateDataTemplate(object content) {
-		    throw new NotImplementedException();
-	    }
+		public object Convert(object value, Type targetType, object parameter, CultureInfo culture) {
+			throw new NotImplementedException();
+		}
 
-	    public ControlTemplate CreateControlTemplate(object content) {
-		    throw new NotImplementedException();
-	    }
+		public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture) {
+			throw new NotImplementedException();
+		}
 
-    }
+	}
 }